# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ComicCrawler REST API
  version: 0.1.0
  description: Public REST contract for API-only crawl clients.
servers:
  - url: http://127.0.0.1:4100
paths:
...
  /api/challenge-discovery/{id}:
    get:
      summary: Read a human verification handoff job
      description: The namespace is historical; public crawl flow treats this as a verification handoff job.
      parameters:
        - $ref: "#/components/parameters/HandoffId"
      responses:
        "200":
          description: Handoff job
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/HandoffJob"
        "404":
          $ref: "#/components/responses/Error"
  /api/challenge-discovery/{id}/open-external-browser:
    post:
      summary: Open isolated browser for human verification
      parameters:
        - $ref: "#/components/parameters/HandoffId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenVerificationBrowserRequest"
      responses:
        "200":
          description: Updated handoff job
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/HandoffJob"
        "400":
          $ref: "#/components/responses/Error"
  /api/challenge-discovery/{id}/complete-human-verification:
    post:
      summary: Check whether human verification is complete
      parameters:
        - $ref: "#/components/parameters/HandoffId"
      responses:
        "202":
          description: Updated handoff job
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/HandoffJob"
        "400":
          $ref: "#/components/responses/Error"
components:
  parameters:
...
    HandoffId:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
    ApiResponse:
      type: object
      required: [data]
      properties:
        data: {}
        error:
          type: string
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
        data: {}
...
    HandoffJob:
      type: object
      required: [id, normalizedUrl, status]
      properties:
        id:
          type: string
        url:
          type: string
        normalizedUrl:
          type: string
        hostname:
          type: string
        status:
          type: string
        error:
          type: string
    OpenVerificationBrowserRequest:
      type: object
      properties:
        executablePath:
          type: string
        profileId:
          type: string